% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_plot.R
\name{acc_plot}
\alias{acc_plot}
\title{Plot Model Accuracy}
\usage{
acc_plot(accuracy_data, y = "correct")
}
\arguments{
\item{accuracy_data}{data generated by mcguiR::accuracy.log()}

\item{y}{desired value to be plotted - could be "correct", "incorrect", or "both"}
}
\value{
a plot of accuracy at each probability threshold
}
\description{
only supports binary logistic regression at this time
}
\examples{
\dontrun{
iris2 <- iris[stringr::str_detect(Species, "setosa", negate = T), ]
irismodel <- glm(Species ~ ., data = iris2, family = binomial)
acc_data <- accuracy(irismodel, iris2, iris2$Species, "virginica", "versicolor")
acc_plot(acc_data, y = "both")}
}
