% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.lm.R
\name{accuracy.lm}
\alias{accuracy.lm}
\title{Determine Accuracy of Linear Models (via differences between actual and
predicted values)}
\usage{
accuracy.lm(model, data, ind_column, round = 2, output = "data")
}
\arguments{
\item{model}{model to be examined}

\item{data}{data to be used in model - MUST include all predictors found in model}

\item{ind_column}{independent (outcome) variable column in data - please use data$column syntax}

\item{round}{number of digits the predicted value to be rounded to}

\item{output}{default is output = "data", which returns a data frame.  setting this
to "plot" will plot the distribution of differences}
}
\value{
data frame or plot
}
\description{
Returns either a data frame of the actual value, the predicted value from the model,
and the difference between the two OR a plot showing the distribution of differences
}
\examples{
\dontrun{
model <- lm(Petal.Length ~ ., data = iris)
accuracy.lm(model, iris, iris$Petal.Length, round = 2, output = "plot")}
}
