% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Title}
\usage{
accuracy(model, data, ind_column, success, failure, na.rm = T)
}
\arguments{
\item{model}{regression model to be used}

\item{data}{training/testing data to use with model, MUST include all predictors in model}

\item{ind_column}{independent (outcome) variable column in data - please use data$column syntax}

\item{success}{character vector; how 'success' is coded in data}

\item{failure}{character vector; how 'failure' is coded in data}

\item{na.rm}{when T, only returns Probability Thresholds that have both 'correct' and 'incorrect' values}
}
\value{
data frame with \% correct calls ('correct'), \% incorrect calls ('incorrect') and the probability threshold
}
\description{
Title
}
