% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_wrapper.R
\name{AUC_wrapper}
\alias{AUC_wrapper}
\title{Wrapper for DescTools::AUC}
\usage{
AUC_wrapper(ROC_values)
}
\arguments{
\item{ROC_values}{data frame generated by mcguiR::ROC_values()}
}
\value{
numeric value
}
\description{
Wrapper for DescTools::AUC
}
