% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_plot.R
\name{log_plot}
\alias{log_plot}
\title{Plot continuous predictors against the logit}
\usage{
log_plot(model, data)
}
\arguments{
\item{model}{the model to use}

\item{data}{the data to use, MUST include all the predictors found in model}
}
\value{
plot of predictors vs. logit
}
\description{
Plot continuous predictors against the logit
}
