% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_samples.R
\name{problem_samples}
\alias{problem_samples}
\title{Find Problem Samples within Training Dataset}
\usage{
problem_samples(
  model,
  data,
  k,
  standard = 2,
  student = 2,
  df_fits = 1,
  cooks = 1
)
}
\arguments{
\item{model}{the model to be used}

\item{data}{TRAINING data set}

\item{k}{number of predictors in model}

\item{standard}{cut-off for standardized residuals- samples with values above
abs(standard) will be returned}

\item{student}{cut-off for studentized residuals - samples with values above
abs(standard) will be returned}

\item{df_fits}{cut-off for DFFITS - samples with values above
abs(standard) will be returned}

\item{cooks}{cut-off for cook's distance - samples with values above
abs(standard) will be returned}
}
\value{
a data frame
}
\description{
only supports binary logistic regression at this time.
doesn't work when dfbeta(model) is inclueded - recomment still looking at
these values outside of this function
}
\examples{
\dontrun{
iris2 <- iris[stringr::str_detect(Species, "setosa", negate = T), ]
irismodel <- glm(Species ~ ., data = iris2, family = binomial)
problem_samples(irismodel, iris2, k = 4)}
}
