% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplot.lm.R
\name{quickplot.lm}
\alias{quickplot.lm}
\title{Plot the True Values vs. the Predicted Values}
\usage{
quickplot.lm(model, data, ind_column)
}
\arguments{
\item{model}{the model to be used}

\item{data}{the data to be fed into the model - MUST include all the predictors
in the model}

\item{ind_column}{independent (outcome) variable column in dataset - please use data$column syntax}
}
\value{
plot
}
\description{
Plot the True Values vs. the Predicted Values
}
\examples{
\dontrun{
model <- lm(Petal.Length ~ ., data = iris)
quickplot.lm(model, iris, iris$Petal.Length)}
}
