% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplot.R
\name{quickplot}
\alias{quickplot}
\title{Plot Samples by their Predicted Probabilities}
\usage{
quickplot(model, data, ind_column, title)
}
\arguments{
\item{model}{model to be used}

\item{data}{testing/training/validation data set, MUST have all the predictors in the model}

\item{ind_column}{independent (outcome) variable column in dataset - please use data$column syntax}

\item{title}{title for the plot}
}
\value{
a plot
}
\description{
only supports binary logistic regression at this time.
}
\examples{
\dontrun{
iris2 <- iris[stringr::str_detect(Species, "setosa", negate = T), ]
irismodel <- glm(Species ~ ., data = iris2, family = binomial)
quickplot(irismodel, iris2, iris2$Species, "Plot")}
}
