% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_plot.R
\name{ROC_plot}
\alias{ROC_plot}
\title{Plot Data Generated by mcguiR::ROC_value()}
\usage{
ROC_plot(ROC_data, method = "loess", title = "")
}
\arguments{
\item{ROC_data}{data frame generated by ROC_value()}

\item{method}{method for geom_smooth to use ("lm", "glm", "gam", "loess") - "loess" is default}

\item{title}{character vector containing preferred title (If left empty, blank)}
}
\value{
plot with FPR on the x axis and TPR on the y axis
}
\description{
Plot Data Generated by mcguiR::ROC_value()
}
