% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_value.R
\name{ROC_value}
\alias{ROC_value}
\title{Calculate TPR and FPR at Each Probability Threshold}
\usage{
ROC_value(model, data, ind_column, success, failure, na.rm = T)
}
\arguments{
\item{model}{regression model to be used}

\item{data}{training/testing data to use with model, MUST include all predictors in model}

\item{ind_column}{independent (outcome) variable column in dataset - please use data$column syntax}

\item{success}{character vector; how 'success' is coded in data}

\item{failure}{character vector; how 'failure' is coded in data}

\item{na.rm}{when T, only returns probability thresholds that have values for both 'TPR' and 'FPR' values}
}
\value{
data frame with TPR, FPR and Probability Threshold columns
}
\description{
Only supports binary logistic regression at this time
}
