% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_plot.R
\name{acc_plot}
\alias{acc_plot}
\title{Plot Model Accuracy}
\usage{
acc_plot(accuracy_data, y = "correct")
}
\arguments{
\item{accuracy_data}{data generated by mcguiR::accuracy()}

\item{y}{desired value to be plotted - could be "correct", "incorrect", or "both"}
}
\value{
a plot of accuracy at each probability threshold
}
\description{
Plot Model Accuracy
}
