% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_wrapper.R
\name{AUC_wrapper}
\alias{AUC_wrapper}
\title{Wrapper for DescTools::AUC}
\usage{
AUC_wrapper(ROC_value)
}
\arguments{
\item{ROC_value}{data frame generated by mcguiR::ROC_values()}
}
\value{
numeric value
}
\description{
Wrapper for DescTools::AUC
}
\examples{
\dontrun{
iris2 <- iris[stringr::str_detect(Species, "setosa", negate = T), ]
irismodel <- glm(Species ~ ., data = iris2, family = binomial)
roc_values <- ROC_value(irismodel, iris2, iris2$Species, "virginica", "versicolor")
AUC_wrapped(roc_values)}
}
